+++ src/vs/workbench/contrib/chat/browser/promptSyntax/attachInstructionsAction.ts
